feat(admin): extract Recess Gear admin configuration to dedicated panel
